+++ src/bibx/algorithms/sap.py
@@ -4,8 +4,8 @@ from typing import Any, cast
 import networkx as nx
 from networkx.algorithms.community.louvain import louvain_communities
 
-from bibx.article import Article
-from bibx.collection import Collection
+from bibx.models.article import Article
+from bibx.models.collection import Collection
 
 YEAR = "year"
 LEAF = "leaf"
