+++ src/bibx/models/article.py
