+++ src/bibx/models/collection.py
