+++ src/bibx/sources/openalex.py
@@ -4,9 +4,9 @@ from enum import Enum
 from typing import Optional
 from urllib.parse import urlparse
 
-from bibx.article import Article
 from bibx.clients.openalex import OpenAlexClient, Work
-from bibx.collection import Collection
+from bibx.models.article import Article
+from bibx.models.collection import Collection
 
 from .base import Source
