+++ src/bibx/sources/scopus_bib.py
@@ -6,9 +6,9 @@ from typing import Optional, TextIO
 
 import bibtexparser
 
-from bibx.article import Article
-from bibx.collection import Collection
 from bibx.exceptions import MissingCriticalInformationError
+from bibx.models.article import Article
+from bibx.models.collection import Collection
 
 from .base import Source
