+++ src/bibx/sources/wos.py
@@ -7,13 +7,13 @@ from contextlib import suppress
 from dataclasses import dataclass
 from typing import Any, Callable, ClassVar, Optional, TextIO, Union
 
-from bibx.article import Article
-from bibx.collection import Collection
 from bibx.exceptions import (
     InvalidIsiLineError,
     InvalidIsiReferenceError,
     MissingCriticalInformationError,
 )
+from bibx.models.article import Article
+from bibx.models.collection import Collection
 
 from .base import Source
